requirements.txt
pandas
openpyxl
sentence-transformers
faiss-cpu
numpy
ollama